tweaks to make numpy work

+++ tracker/vball_sort.py
@@ -34,7 +34,7 @@ class STrack(BaseTrack):
     def __init__(self, tlwh, score, cls, jumping, feat=None, feat_history=50):
 
         # wait activate
-        self._tlwh = np.asarray(tlwh, dtype=np.float)
+        self._tlwh = np.asarray(tlwh, dtype=np.float32)
         self.kalman_filter = None
         self.mean, self.covariance = None, None
         self.is_activated = False
@@ -213,7 +213,7 @@ class STrack(BaseTrack):
         """Get delta x and delta y
         """
         if self.mean is None:
-            return np.asarray([0, 0], dtype=np.float)
+            return np.asarray([0, 0], dtype=np.float32)
         return self.mean[4:6].copy()
 
     @property
